Correção Arquivos 1.0
Correção arquivos CRUD.

+++ TrabalhoJavaCrud-WilsonAlzuguir/src/main/webapp/lista_curso.jsp
@@ -3,6 +3,7 @@
 <%@ page import="br.com.cursoja.controlecursoja.model.dao.CursoDao" %>
 <%@ page import="br.com.cursoja.controlecursoja.model.entidade.Curso" %>
 <%@ page import="java.util.List" %>
+<%@ page import="java.text.DecimalFormat" %>
 <!DOCTYPE html>
 <html>
 <head>
@@ -10,7 +11,20 @@
 <title>Nossos Cursos</title>
 </head>
 <body>
+	<jsp:include page="verificaSessao.jsp"></jsp:include>
+	<a href="area_interna.jsp">Voltar a página anterior</a>
+	<a href="incluir_curso.jsp">Cadastro de Cursos</a>
 	<h1>Todos os Cursos</h1>
+	<%
+		String nomeBusca = "";
+		if(request.getParameter("nomeBusca")!= null){
+			nomeBusca = request.getParameter("nomeBusca");
+		}
+	%>
+	<form method="post" action="lista_curso.jsp">
+    	<input type="text" name="nomeBusca" value="<%=nomeBusca%>"/>
+    	<input type="submit" value="Buscar"/>
+	</form>
 	<table>
 		<thead>
 			<tr>
@@ -22,15 +36,18 @@
 		<tbody>
 		<%
 			CursoDao dao = new CursoDao();
-			List<Curso> cursos = dao.listar("");
+			List<Curso> cursos = dao.listar(nomeBusca);
+			DecimalFormat fmt = new DecimalFormat("###, ##0.00");
+			String strValor = "";
 			for (Curso c: cursos){
+				strValor = fmt.format(c.getValor());
 		%>
 			<tr>
 				<td><%= c.getNome() %></td>
-				<td><%= c.getValor() %></td>
+				<td><%= strValor %></td>
 				<td>
 					<a href="IniciarAlterarCurso?id=<%= c.getId()%>">Alterar</a> 
-					<a href="ExcluirCurso?id=<%= c.getId()%>">Excluir</a>
+					<a href="ExcluirCurso?id=<%= c.getId()%>" onclick="return confirm('Deseja realmente excluir esse registro?');">Excluir</a>
 				</td>
 			</tr>
 			<% } %>
